+++ src/App/Committees/Compucie/index.tsx
@@ -3,12 +3,12 @@ import {useDispatch, useSelector} from "react-redux";
 import {chwazi, fetchInitialData} from "actions";
 import {compucieSelector} from "./../selectors";
 import Compucie from "./Compucie";
-import {useProductPurchase} from "App/Products/Context";
+import {useOrder} from "App/Products/OrdersContext";
 
 const CompucieScreen = () => {
   const dispatch = useDispatch();
   const {compucie, scriptcie} = useSelector(compucieSelector);
-  const {selectMember} = useProductPurchase();
+  const {selectMember} = useOrder();
 
   return (
     <Compucie
