+++ src/App/Prominent/index.tsx
@@ -1,12 +1,13 @@
 import {useSelector} from "react-redux";
 import {boardsSelector, prominentSelector} from "./selectors";
 import Prominent from "./Prominent";
-import {useProductPurchase} from "App/Products/Context";
+import {useOrder} from "App/Products/OrdersContext";
 
 const ProminentScreen = () => {
   const boards = useSelector((state: any) => boardsSelector(state));
   const prominent = useSelector((state: any) => prominentSelector(state));
-  const {selectMember} = useProductPurchase();
+  const {selectMember} = useOrder();
+
   return <Prominent boards={boards} prominent={prominent} selectMember={selectMember} />;
 };
