+++ src/actions.ts
@@ -57,10 +57,6 @@ export const TYPES = {
   FETCH_ACTIVITIES_REQUEST: "FETCH_ACTIVITIES_REQUEST",
   FETCH_ACTIVITIES_SUCCESS: "FETCH_ACTIVITIES_SUCCESS",
   FETCH_ACTIVITIES_FAILURE: "FETCH_ACTIVITIES_FAILURE",
-
-  AUTHENTICATE_REQUEST: "AUTHENTICATE_REQUEST",
-  AUTHENTICATE_SUCCESS: "AUTHENTICATE_SUCCESS",
-  AUTHENTICATE_FAILURE: "AUTHENTICATE_FAILURE",
 };
 
 const orderQueue = {};
